add feature to show other users and distribute views as partials

+++ controllers/comment_controller.js
@@ -26,16 +26,18 @@ module.exports.destroyComment = (req, res) => {
         try{
             const comment = await Comment.findById(req.query.commentId);
             if(comment){
-                const post = await Post.findById(req.query.postId);
-                post.comment = post.comment.filter((id) => {
-                    return id != req.query.commentId;
-                });
-                post.save();
-                await Comment.deleteOne({_id: req.query.commentId});
+                if(comment.user == req.user.id){
+                    const post = await Post.findById(req.query.postId);
+                    post.comment = post.comment.filter((id) => {
+                        return id != req.query.commentId;
+                    });
+                    post.save();
+                    await Comment.deleteOne({_id: req.query.commentId});
+                }
             }
             return res.redirect('back');
         }catch(err){
-            console.log(`Error: ${err}`);
+            console.log(`Error: ${err.message}`);
         }
     })();
 }
